remove common context
