+++ src/core/internalMonitoring.ts
@@ -4,9 +4,7 @@ import lodashAssign from 'lodash.assign'
 
 import { computeStackTrace } from '../tracekit/tracekit'
 import { Configuration } from './configuration'
-import { getCommonContext } from './context'
 import { toStackTraceString } from './errorCollection'
-import { Session } from './session'
 import { StatusType } from './status'
 import { Batch, HttpRequest } from './transport'
 import * as utils from './utils'
@@ -40,10 +38,10 @@ export function startInternalMonitoring(configuration: Configuration) {
     configuration.maxMessageSize,
     configuration.flushTimeout,
     () => ({
+      date: new Date().getTime(),
       http: {
         referer: window.location.href,
       },
-      ...getCommonContext(),
     }),
     utils.withSnakeCaseKeys
   )
