+++ src/core/utils.ts
@@ -1,5 +1,3 @@
-import { Context, ContextValue } from './context'
-
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
 
 export const ONE_SECOND = 1000
@@ -70,6 +68,14 @@ export function round(num: number, decimals: 0 | 1 | 2 | 3) {
 export function msToNs(duration: number) {
   return round(duration * 1e6, 0)
 }
+
+export interface Context {
+  [x: string]: ContextValue
+}
+
+export type ContextValue = string | number | boolean | Context | ContextArray | undefined
+
+export interface ContextArray extends Array<ContextValue> {}
 
 export function withSnakeCaseKeys(candidate: Context): Context {
   const result: Context = {}
