+++ test/e2e/scenario/helpers.ts
@@ -1,11 +1,9 @@
 import * as request from 'request'
-import { CommonContext } from '../../../src/core/context'
 import { ErrorContext, HttpContext } from '../../../src/core/errorCollection'
 import { MonitoringMessage } from '../../../src/core/internalMonitoring'
 import { LogsMessage } from '../../../src/logs/logger'
 import { RumEvent } from '../../../src/rum/rum'
 
-export type ServerRumEvent = RumEvent & CommonContext
 export interface ServerErrorMessage {
   error: ErrorContext
   http: HttpContext
@@ -37,12 +35,12 @@ export async function tearDown() {
 }
 
 export async function retrieveRumEvents() {
-  return fetch('/rum').then((rumEvents: string) => JSON.parse(rumEvents) as ServerRumEvent[])
+  return fetch('/rum').then((rumEvents: string) => JSON.parse(rumEvents) as RumEvent[])
 }
 
 export async function retrieveRumEventsTypes() {
-  return retrieveRumEvents().then((rumEvents: ServerRumEvent[]) =>
-    rumEvents.map((rumEvent: ServerRumEvent) => rumEvent.evt.category)
+  return retrieveRumEvents().then((rumEvents: RumEvent[]) =>
+    rumEvents.map((rumEvent: RumEvent) => rumEvent.evt.category)
   )
 }
